feat: rework design a bit + mobile design
